requirements.txt
Django==1.11.7
django-haystack==2.6.1
django-pure-pagination==0.3.0
django-taggit==0.22.1
jieba==0.39
Markdown==2.6.10
mysqlclient==1.3.12
olefile==0.44
Pillow==4.3.0
Pygments==2.2.0
pytz==2017.3
Whoosh==2.7.4
